+++ packages/translation-nmtcpp/examples/quickstart.js
@@ -65,7 +65,7 @@ async function testBergamot() {
       srcVocab: path.join(modelDir, fileNames.srcVocabName),
       dstVocab: path.join(modelDir, fileNames.dstVocabName)
     },
-    params: { mode: 'full', dstLang, srcLang },
+    params: { dstLang, srcLang },
     config: {
       modelType: TranslationNmtcpp.ModelTypes.Bergamot
     },